backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-aws
boto3
python-pptx